chore: refining account codebase
