+++ pyproject.toml
@@ -20,6 +20,9 @@ python-dotenv = "^1.0.0"
 # workaround for issue with copier dependency spec allowing major upgrade to pydantic v2
 pydantic = "^1.10.2"
 mslex = "^1.1.0"
+keyring = "^24.2.0"
+pyjwt = "^2.8.0"
+auth0-python = "^4.4.0"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.2.0"
